RHI: Fix crash in QQuickWindow when updating render target
The crash is reproducible e.g. when a resize event comes
after a render target has been created, and it has to be
recreated. It would then delete the render pass descriptor
but it would still be in the pipeline cache in the renderer.
We need to tell the renderer it is invalidated before deleting
it.

Discovered as part of making QQuickWidget work on RHI. The
crash was reproducible by creating a new tab widget in the
qquickwidget example.

Task-number: QTBUG-78638
Change-Id: Ia840cc423a811aee18c568d43624b6bf852246f7
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>