Massage toolTipCrashOnClose test
QTest::ignoreMessage() has a feature, namely that:
  If the test finished and the message was not outputted,
  a test failure is appended to the test log.

So here, if Graphical Effects is available, but not functional,
ShaderEffect fails to initialize, which means the warning-to-be-skipped
is never outputted. The test would pass (because the visual appearance
does not matter), but the ignoreMessage() condition fails it.

Adding a QSKIP is not a solution because ignoreMessage() would still
do its check and fail the test! Therefore, remove the ignoreMessage()
altogether.

Change-Id: I3e096752dfc6894c77859e786c8a4cac01d6fc53
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>