JIT: Don't pass the heap object in the return value register
We have two variants for call destination results: Ignore or store in
accumulator. The accumulator may be the same as the return value
register. Therefore, when side stepping this mechanism and leaving the
result in the return value register, we should have realized that we
have to save/load the accumulator. We didn't and left the accumulator
clobbered.

A further fix actually clears the accumulator proactively after each
call with Ignore if it is the same as the return value register. Now we
don't get our result anymore in this case.

To fix this, just use the accumulator register, declare that we do so,
and save/load it around the call.

Change-Id: I65fe4f7430ebfc419863b329bcd66a95875e5ffb
Reviewed-by: Lars Knoll <lars.knoll@qt.io>