Enable invoking the scenegraph rendering without a full render pass
Split the monolithic render() and renderBatches() into four steps
(prepare, begin-pass, render, end-pass), of which two are optional.

The behavior of the batch renderer's render() (the entry point when
rendering a Qt Quick scene normally) remains unchanged: it just calls
prepare, beginPass, render, endPass which together should boil down to
exactly what the monolithic implementation was doing before.

However, we can now also support recording rendering of the 2D scene
as part of a render pass that is started by something else (e.g. the 3D
scene). To enable this, the renderer interface gets prepareInline() and
renderInline() virtuals that are only implemented by the batch renderer
(and are no-ops by default).

These are complemented by prepareSceneLine() and renderSceneInline()
which are public (as in the class, not API-wise), like renderScene().
The former does preprocess() and triggers the internal prepareInline(),
whereas renderSceneLine() is where renderInline() is called from. The
rendercontext has no corresponding wrappers since one can just directly
call these on the QSGRenderer.

Change-Id: Iddfcec6970bad2bc73e6e8af19fe95d16c6ac9e5
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>