Increase the timeout for tst_qmlmin
This test has been flaky during CI runs (on macOS) because sometimes
it takes more time to execute than the default 5 minutes limit.

Probably not the ideal fix, but since this will be deprecated for Qt 6
there is no point in investing too much time into fixing it.

Task-number: QTBUG-84339
Change-Id: Iab53a05a68c32a1fae56cafdeeb6aa74900ac9fc
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>