Fix offset on native text rendering with high-dpi scaling
The glyph positions are scaled *after* they have been positioned,
so when we subtracted the margin unmodified, we would actually
offset by the scaled margin (so for scale factor=2 it would be
2 pixels off, and for scale factor=4 it would be 6 pixels off,
etc.)

We have to also prescale the margin by the inverse of the scale
we will apply later.

[ChangeLog][Text] Fixed an offset on text position when combining
NativeRendering with high-dpi scaling.

Task-number: QTBUG-84454
Pick-to: 5.15
Change-Id: I703aeb7fbd717bee5d88cc61e9a56c6422558889
Reviewed-by: Lars Knoll <lars.knoll@qt.io>