Fix a clipping bug in software-rendered text
ac179e235ba0c01fff6dd5f4ad2cc9696fe78822 fixed a clipping issue with
software rendering and text by using the actual bounding rect of the
alpha map (which may be extended to allow for margins around the glyph).

But it did not account for the margins on the top and left sides of the
glyph, causing QTBUG-84042.

By a coincidence, this issue was not reproducible with the DirectWrite
engine in Qt 5.15 at the time because of 318a991907b6c08f52786160bafea1e30d3ad9bd
in Qt Base, which baked the margins into the top and left positions of the
returned bounding rect. But this is not what the other font engines are doing,
and caused issues where the margin would be accounted for twice in the output.
So 318a991907b6c08f52786160bafea1e30d3ad9bd was reverted.

This change shifts the bounding rect by the margins (if any), avoiding clipping
on the left and top sides.

Task-number: QTBUG-80180
Fixes: QTBUG-84042
Pick-to: 5.15
Change-Id: I942f05f0e8c8eb8c5b3071a73406e3e744b7d5a0
Reviewed-by: Andy Nichols <andy.nichols@qt.io>