Sync the QQuickPlatformMenu so it has a handle ready when its a sub menu
When dynamically creating a menu, it will be parented after it has
created the handle the first time around. Therefore it loses the original
handle and does not get a new one when needed. So by calling sync()
before it is set as the sub menu then it can get a handle created if
necessary in time.

Change-Id: Ia39f49f99758802dd19ff1df478b05ac5e403ea8
Pick-to: 5.15
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>