Native style: improve type safety
A QuickStyleItem instantiated inside a QML file set property "control" to
point to the control it styles. It's required that this item matches the
control it styles, so that e.g a NativeStyle.Button is only used for styling
a Button.qml (and not e.g a Slider.qml), since NativeStyle.Button
will crash otherwise. We do a quick runtime check in debug mode rather than
a compile time check to avoid to much churn in the c++ code to enforce this.

Change-Id: I31dca31ab4f1fdf08780ffb01c533ac5837784f2
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>