Native style: remove insets, and use layoutRect
We have so far used insets to ensure that the controls
end up aligned inside a layout. But this is most likely wrong, since
insets in QtQuickControls2 is supposed to only move the
background, and not the whole control (including contents,
which we used to move as well in Button.qml). So after looking at
this one more time, I think the right solution is to leave
insets alone, and instead provide a layoutRect from
QQuickStyleItem. This rect can later be used by layouts to
align the controls correctly.

Change-Id: I3cd97d34ca6f629dedef3e2afd2ae9e257f9df22
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>