Native style: let DefaultTextField have correct contentRect and size
Set implicitWidth/Height to be the same as in TextField.qml in the
default style (and the same for TextArea). And fix the mac style to
return correct values when calculating the minimum size. Combined, this
should ensure that we improve how we take the content size into account
when calculting the implicit size of a TextField (especially on macOS).

Change-Id: I39bc8bb0f238dd29d5fe33fc32d888b82d020b0e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>