Native style: don't clean geometry flag
We used to only repaint the image upon getting
a geometryChanged callback, to avoid recursion.
But this turns out to fail when the resize is
a result of e.g a change in the content size.
In that case, we need to recalculate geometry
as well, since many of the rects are calculated
based on the content size.

This patch will change this logic so that we
always recalculate style geometry upon a item
geometry change, unless we're already polishing
(that latter would otherwise produce a warning).

Change-Id: I06c2e527d03dca54c555ddb8ab5d9b2a9546163d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>