Make Binding restoreMode qmlWarning follow a logging category
Binding{} emits a warning if restoreMode is not specified

Unfortunately that syntax is only available in Qt >= 5.14.
Projects need to support Qt 5.15 and also older (5.12), which means we have code constantly hitting
this warning. We can't even #ifdef the QML code.

By introducing a logging category we can still have the warning but
provides some option for downstream to temporarily disable through an
existing mechanism until older Qt is not supported.

Change-Id: I47190b7216b776d9db6fbece63332e993fce7203
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>