QQuickStackView: deprecate Transition enum value
bc86a96c56934d98d641844319a905753f56895a added a comment that this was
deprecated, but didn't document it as deprecated or have any mention of
it being deprecated besides that one comment.

As we are dynamically evaluating the JS arguments in C++, we don't need
to have a "default" enum value; we can detect omission of the argument
and act accordingly.

It's now too late to deprecate it for Qt 5, but we can do it for Qt 6.

[ChangeLog][Controls][StackView] The StackView.Transition enum value
was deprecated. The operation argument can be omitted in order to use
the default transition for any given operation.

Task-number: QTBUG-84715
Change-Id: I530235beb2ebf56deb6f82ab54e269322e92a0a5
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>