Prevent premature child destruction
QQmlContextData::emitDestruction suffers from the fact that code can
delete objects while emitDestruction is ongoing. Notably, the sequence
child->emitDestruction can trigger a call to a->destruction (of one of
child's attached components), which then can indirectly delete both
child and child->nextChild (for instance, when a StackView gets
cleared).
We prevent this by using QQmlRefPointer when iterating over the
children, which keeps the child alive for the duration of the loop.
This solves the same issue as 0c8e51705ac0bb86c4b123ecd30a11b41fd50b24
in 5.15

Fixes: QTBUG-84095
Change-Id: I486a65f791c2a52d6c5c0f188997225b6db75267
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>