Prevent Vulkan validation warning in 3D mode with text
Setting a dynamic state on the command buffer is an error if the pipeline
was not created with the appropriate flag.

In the depth post pass the pipeline's flags do not reflect the modifications
from the materials' updateGraphicsPipeline(). So we should not set the
appropriate dynamic state either. (blending has no relevance for the depth
output anyway)

Change-Id: Ifedae1f5175ac4f1905dec4b7a5c305417234808
Reviewed-by: Andy Nichols <andy.nichols@qt.io>