Remove quality evaluation in test string
It is not the test's place to evaluate the quality of the outline
shader (which also looks perfectly fine unless you don't really
like outlined text). And it looks unprofessional to use rude words
to describe our code in manual tests. Strings in lancelot tests are
also inefficient ways of reporting and tracking bugs if that was the
intended use of this.

Change-Id: Ib9fdb4e56539b9b02f8b3f1d06eb3e3beb51f8ef
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit facf085fbcf614a25c6cfdff39ad80a46180a322)
Change-Id: Icc62114702265e9e1b247624370a78999e564d5f
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>