Fix QQmlPropertyBinding::evaluate
When a class is in a semi-deleted state (QQmlData::wasDeletet returns
true, but the QQmlElement has not been removed yet), we can still end up
calling evaluate under some circumstance. This can be observed in qqc2's
tests, for instance tst_palette. In those cases either ctxt or
ctxt->engine() might be nullptr. We handle this by returning an error;
as the object is about to be gone, the result does not matter anyway.

Change-Id: I8a42263a9b44012822f5a827da8eae97d3160165
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>