Native style: offer layoutMargins to QML, not layoutRect
The layoutRect are, similar to contentRect, calculated based
on the minimumSize of the control. Since we don't know what
the actual layoutRect will be until a control is resized, we
use it instead to calculate what the margins should be between
the boundingRect() and layoutRect. We do the same for content
rect. This will allows us to calculate the correct margins
early on, before knowing what the end size of the control will be.

Change-Id: I26b0aedee28533a70fd8dc4918f962f41c4f5185
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>