import Qt namespace in QtQml
Introduces and uses the QML_FOREIGN_NAMESPACE to declare the Qt namespace in QtQml, and exports all the enumerations
in the Qt namespace, and makes them available also in QtQml/plugins.qmltypes

Please note that this does *not* remove the special QV4::QtObject
as it contains much more than the enumerations.
Having two object registered as Qt (the complete QV4::QtObject added
to the global properties in ExecutionEngine::initializeGlobal, and
QtInQml registered by the qmltyperegistrar) gives problems, if the
partial one overrides the complete one.

For this reason the classinfo QML.ManualRegistration (no public macro
until shown to be more generally useful) is introduced and
used to ensure that QtInQml generates only documentation
(plugins.qmltypes), and no registration that might conflict with
QtObject.

This makes the current builtins.qmltypes redundant.

Fixes: QTBUG-84897
Change-Id: Iffc96cd3e1d9f9d7d047ee60221a4b0928f53005
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>