Make sure the focus rect is drawn
In windows style, the focus rect is drawn when the button is drawn,
but we are only drawing the bevel, so the focus rect was not drawn.

We move the logic of drawing the focus rect so that it is drawn when
drawing the bevel instead.

Also, we have to set the QStyle::State_KeyboardFocusChange in order
for the focus rect to be drawn

Change-Id: I4463ae1dd0f23ecc5bb0a84c563fda33dc6e93c8
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>