QQuickWindow: Remove (set)ClearBeforeRendering
The doc update reveals that this setting has been ignored with the RHI
always, and there is no plans (or ways) to support this.

As the beforeRenderPassRecording() signal documents, connecting to that
signal is effectively an alternative to disabling clearing the color
buffer. (because it allows the application to execute custom GL or other
calls after the clear but before the rest of Qt Quick's rendering)

Therefore, remove the clearBeforeRendering flag from Qt 6.

[ChangeLog][Qt Quick][QQuickWindow] setClearBeforeRendering() and
clearBeforeRendering() are removed.

Change-Id: I6a0a6f56552331923f2e07390e2f33f58c6c31e0
Reviewed-by: Andy Nichols <andy.nichols@qt.io>