Update ShaderEffect docs
The two ShaderEffect related snippets are removed, with the relevant
code folded into the documentation itself. We cannot support standalone
snippets with custom shaders as we'd need to have separate shader files,
qsb files, scripts to build, etc. This is not reasonable at this point.

In layerblending.qml the ShaderEffect serves no purpose, apart from
providing a fancy background that is irrelevant to the topic. Just
remove that.

Change-Id: I726f029c8b7096938aa609031e391e9e57ef1655
Reviewed-by: Christian Strømme <christian.stromme@qt.io>