Simplify the software backend selection logic
Avoid relying on the (somewhat obsolete) capability flag. Instead, have
a clean and straightforward check based on QT_CONFIG.

In the CI, in particular for the QEMU configs that are both OpenGL and
Vulkan capable (build-wise, yuck), the software backend will be requested
explicitly in the future so this won't need to cater for that anymore.

Change-Id: Ifa100efdae208b789f6ed2a7b57330262c13284f
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>