Remove broken profiler/trace points in the scenegraph
As the todo note says: this profiling is all wrong

Not reporting something is better than reporting something bogus.

The other profiling points, f.ex. in the render loops, seem to work as
expected and give reasonable results in the QML Profiler view in Creator.

The autotest needs fixing because there is no SGContextFrame events left
with the rhi code path either atm, just like with the software backend.

Change-Id: Iee2d3feb586108db50c83c6c03ea415384561ede
Reviewed-by: Christian Strømme <christian.stromme@qt.io>