Move resetOpenGLState out of QQFBO
QQFBO is not the only client of resetOpenGLState. Although in theory
third-party GL code that integrates with QtQuick should reset its
state, in practice it doesn't. By making resetOpenGLState
only available into QQFBO, now we're blocking a Qt5->6 upgrade path.

There's also no compelling reason for this function to be in QQFBO
at all, so move it out to as a free function in a ad-hoc new namespace.

Change-Id: Ic8e5c7e244db37a5b6257d516e6aea3a9db44898
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>