Fix outline shader with transparent fill
A bug snuck in via e8d9bc1bcada7d94af5a33d64a9afc860ede5b84 in
one of the outline shaders where we would actually fill the
entire glyph with the style color instead of just drawing the
border. Since we were drawing the fill on top afterwards, this
wasn't visible in a normal test, but when the fill color is
set to transparent it showed.

This changes the alpha channel-based outline shader to match
the old code as well as the code in the red-channel based one.

Task-number: QTBUG-69301
Change-Id: I2f9d4b9426577c331866172fc6f7240f379efef2
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>