Add QQuickHandlerPoint::device property
The HandlerPoint type is used as the memory of an event or TouchPoint
that a handler has already handled.  Maybe in the future we'll store
a copy of a QEventPoint instead.  But for now, it's nice to have the
device pointer available for binding properties, instead of only in
QQuickEventPoint, which was only exposed in signals like
TapHandler.tapped().

Change-Id: I314fc8c76311dc2620f1d97d4cadacf2e9869274
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>